+++ choice-program.el
@@ -7,7 +7,7 @@
 ;; Maintainer: Paul Landes
 ;; Keywords: execution processes unix lisp
 ;; URL: https://github.com/plandes/choice-program
-;; Package-Requires: ((emacs "29") (dash "2.17.0"))
+;; Package-Requires: ((emacs "27.1") (dash "2.19.1"))
 
 ;; This file is not part of GNU Emacs.